% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{class_timetable}
\alias{class_timetable}
\title{Class timetable}
\format{
A data frame with 10 rows and 9 variables:

\describe{
  \item{school_id}{ID of school}
  \item{class_id}{ID of class}
  \item{date}{Date}
  \item{school_start}{School start time}
  \item{school_end}{School end time}
  \item{recess1_start}{First recess period start time}
  \item{recess1_end}{First recess period end time}
  \item{recess2_start}{Second recess period start time}
  \item{recess2_end}{Second recess period end time}
}
}
\usage{
data(class_timetable)
}
\description{
A dataset containing
}
\keyword{datasets}
