% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{home}
\alias{home}
\title{Home points}
\format{
A data frame with 2 rows and 3 variables:

\describe{
  \item{identifier}{Participant identifier}
  \item{home_id}{ID of home}
  \item{geometry}{Point geometry}
}
}
\usage{
data(home)
}
\description{
A dataset containing
}
\keyword{datasets}
