% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{palms}
\alias{palms}
\title{PALMS data}
\format{
A data frame with 40233 rows and 13 variables:

\describe{
  \item{identifier}{Participant identifier}
  \item{datetime}{Timestamp}
  \item{dow}{}
  \item{fixtypecode}{}
  \item{iov}{}
  \item{tripnumber}{}
  \item{triptype}{}
  \item{tripmot}{}
  \item{activity}{}
  \item{activityintensity}{}
  \item{activityboutnumber}{}
  \item{sedentaryboutnumber}{}
  \item{geometry}{}
}
}
\usage{
data(palms)
}
\description{
A dataset containing
}
\keyword{datasets}
