% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{palms_add_domain}
\alias{palms_add_domain}
\title{Add an aggregation domain to the palmsplus_domains table.}
\usage{
palms_add_domain(name, formula)
}
\arguments{
\item{name}{The name of the domain as a string.}

\item{formula}{The formula as a string. This should evaluate to a boolean.
The formula can contain variables already created with \code{palms_add_field}
or other previously specified domains.

Make sure to escape quotation marks that are part of the formula.}
}
\value{
If the palmsplus_domains table is not present in the global environment,
it will be created. If it already exists, the new domain will be appended.
}
\description{
Add an aggregation domain to the palmsplus_domains table.
}
\examples{
palms_add_domain("home", "at_home")

}
