% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{palms_add_multimodal_field}
\alias{palms_add_multimodal_field}
\title{A a multimodal field to the multimodal_fields table}
\usage{
palms_add_multimodal_field(name, func)
}
\arguments{
\item{name}{The name of the field as a string. This field mulst be present in
the \code{trajectories} dataset. This also accepts a vector of names.}

\item{func}{The summary function to use as a string. This is normally either
"sum" or "mean".}
}
\value{
If the multimodal_fields table is not present in the global environment,
it will be created. If it already exists, the new field will be appended.
}
\description{
A a multimodal field to the multimodal_fields table
}
\examples{
palms_add_multimodal_field("mvpa", "sum")
palms_add_multimodal_field("speed", "mean")

# A vector of names using the same summary function
palms_add_multimodal_field(c("sedentary", "light", "moderate", "duration"), "sum")

}
