% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{palms_add_trajectory_field}
\alias{palms_add_trajectory_field}
\title{Adds a trajectory field to the trajectory_fields table.}
\usage{
palms_add_trajectory_field(name, formula, after_conversion = FALSE)
}
\arguments{
\item{name}{The name of the field as a string.}

\item{formula}{The formula as a string.}

\item{after_conversion}{Logical. If TRUE, this field will be calculated after
the trajectory LINESTRING has been created. For example, st_length().}
}
\value{
If the trajectory_fields table is not present in the global environment,
it will be created. If it already exists, the new field will be appended.
}
\description{
Adds a trajectory field to the trajectory_fields table.
}
\examples{
#palms_add_trajectory_field("mot", "first(tripmot)")

}
