% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{palms_add_trajectory_location}
\alias{palms_add_trajectory_location}
\title{Add a trajectory location to the trajectory_locations table.}
\usage{
palms_add_trajectory_location(name, start_criteria, end_criteria)
}
\arguments{
\item{name}{The name of the trip as a string.}

\item{start_criteria}{The start of the trip. This should evaluate to a boolean.
For optimal speed, this should be a field pre-calculated in \code{palmsplus}.}

\item{end_criteria}{The end of the trip. This should evaluate to a boolean.
For optimal speed, this should be a field pre-calculated in \code{palmsplus}.}
}
\value{
If the trajectory_locations table is not present in the global environment,
it will be created. If it already exists, the new location will be appended.
}
\description{
This function builds the trajectory_locations table, which is used
by \code{\link{palms_build_trajectories}} to identify trajectories that start
and end at specific locations. For example, a trip from home to work, or from
school to a park. This is done by setting a start and end criteria.
}
\examples{
# A full example workflow
data(list = c("palms", "home", "school", "participant_basis"))
palms_add_field("at_home", "palms_in_polygon(., filter(palms_buffer(home, distance = 100),
                                                       identifier == i), identifier)")

palms_add_field("at_school", "palms_in_polygon(., filter(school, school_id == participant_basis \%>\%
                                                  filter(identifier == i) \%>\% pull(school_id)))")
palmsplus <- palms_build_palmsplus(palms)

palms_add_trajectory_location("home_school", "at_home", "at_school")

trajectories <- palms_build_trajectories

}
