% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{palms_add_trip_location}
\alias{palms_add_trip_location}
\title{Adds a trip location to the trajectory_locations table.}
\usage{
palms_add_trip_location(name, start_criteria, end_criteria)
}
\arguments{
\item{name}{The name of the field as a string.}

\item{start_criteria}{The start of the trip. This should be a field in
palms plus, and evaluate to a logical.}

\item{end_criteria}{The start of the trip. This should be a field in
palms plus, and evaluate to a logical.}
}
\value{
If the trajectory_locations table is not present in the global environment,
it will be created. If it already exists, the new location will be appended.
}
\description{
Adds a trip location to the trajectory_locations table.
}
\examples{
#palms_add_trip_location("home_school", "at_home", "at_school")

}
