% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/days.R
\name{palms_calc_days}
\alias{palms_calc_days}
\title{Summarises palmsplus data by aggregation domain}
\usage{
palms_calc_days(data)
}
\arguments{
\item{data}{The palmsplus data obtained from \code{palms_calc_plusplus}.}
}
\value{
A table summaised by day. Each day has columns for each domain and
each field specified in \code{palms_add_domain} and \code{palms_add_field}.
}
\description{
Summarises palmsplus data by aggregation domain
}
\examples{
#days <- palms_calc_days(palmsplus)

}
