% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectories.R
\name{palms_calc_trajectories}
\alias{palms_calc_trajectories}
\title{Build trajectories from trip points}
\usage{
palms_calc_trajectories(data)
}
\arguments{
\item{data}{The palms data}
}
\value{
A table of individual trips represented as LINESTRING geometry. Each
trip has a number of fields specified in \code{palms_add_trajectory_field}.
}
\description{
Build trajectories from trip points
}
\examples{
#trajectories <- palms_calc_trajectories(palms)

}
