% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{palms_epoch_length}
\alias{palms_epoch_length}
\title{Returns the epoch length of the PALMS dataset}
\usage{
palms_epoch_length(data)
}
\arguments{
\item{data}{The data points}
}
\value{
An integer representing the seconds between the first and second
data points.
}
\description{
Returns the epoch length of the PALMS dataset
}
\examples{
#epoch <- palms_epoch_length(data)

}
