% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{palms_in_polygon}
\alias{palms_in_polygon}
\title{Returns a logical vector if input points are inside polygons}
\usage{
palms_in_polygon(data, polygons, collapse_var = NULL)
}
\arguments{
\item{data}{The data points.}

\item{polygons}{The polygon feature.}

\item{collapse_var}{An optional parameter for aggragating polygons by a
variable. Essentially this dissolves the polygon layer by \code{collapse_var}.}
}
\value{
A logical vector the length of the input data points. Each polygon
will return an additional column
}
\description{
Returns a logical vector if input points are inside polygons
}
\examples{
#data$is_home <- palms_in_polygon(data, homes, "identifier")

}
