% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{palms_in_time}
\alias{palms_in_time}
\title{Check if points fall between a start and end time}
\usage{
palms_in_time(data, identifier, timetable, basis, start_col, end_col)
}
\arguments{
\item{data}{The data points.}

\item{identifier}{The participant identifier.}

\item{timetable}{The timetable table.}

\item{basis}{The participant basis table.}

\item{start_col}{The column in \code{timetable} containing the start time.}

\item{end_col}{The column in \code{timetable} containing the end time.}
}
\value{
A logical vector the length of the input data points.
}
\description{
Returns a logical vector the length of the input \code{data} points.
This checks if each point falls between a start and end time. This is designed
to work with a \code{timetable} and \code{participant basis} file.
\itemize{
\item \code{timetable} contains a list of start and end times for different periods
of the day, for different groups of people (e.g., classrooms).
\item \code{participant basis} contains a list of identifiers and group membership.
}
}
\examples{
data(list = c("palms", "participant_basis", "class_timetable"))

in_school <- palms_in_time(palms, "BC0627", class_timetable,
                           participant_basis, school_start, school_end)

}
