% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{palms_line_point}
\alias{palms_line_point}
\title{Returns the start or end point of a LINESTRING}
\usage{
palms_line_point(line, first_point)
}
\arguments{
\item{line}{The line geometry}

\item{first_point}{Logical. If true, the first point will be returned.}
}
\value{
A POINT geometry feature representing the first or last point in the
LINESTRING
}
\description{
Returns the start or end point of a LINESTRING
}
\examples{
#start <- palms_line_point(line, TRUE)
#end <- palms_line_point(line, FALSE)

}
