% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{palms_load_defaults}
\alias{palms_load_defaults}
\title{Populates field and domain tables with default values}
\usage{
palms_load_defaults(epoch_length)
}
\arguments{
\item{epoch_length}{The epoch length of the PALMS data in seconds. This is
necessary so the default formulas are created correctly. This can be passed in
using \code{\link{palms_epoch}}.}
}
\description{
This will populate the \code{palmsplus_fields} and
\code{trajectory_fields} with simple fields that should be compatable with
all PALMS data sets. The formulas only use variables that are present in the
PALMS input dataset.
}
\examples{
epoch <- palms_epoch(palms)
epoch

palms_load_defaults(epoch)

}
