% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{palms_remove_tables}
\alias{palms_remove_tables}
\title{Remove all palmsplusr field and domain tables}
\usage{
palms_remove_tables()
}
\description{
Removes all field and domain tables that exist in the global
environment. This is useful when rebuilding tables. This function checks for
the following tables:
\itemize{
\item \code{palmsplus_domains}
\item \code{palmsplus_fields}
\item \code{trajectory_fields}
\item \code{trajectory_locations}
\item \code{multimodal_fields}
}
}
\examples{
palms_remove_tables()

}
