% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{palms_trajectory_point}
\alias{palms_trajectory_point}
\title{Get the start or end point of a LINESTRING}
\usage{
palms_trajectory_point(line, first_point)
}
\arguments{
\item{line}{The \code{LINESTRING} geometry.}

\item{first_point}{Logical. If \code{TRUE}, the first point will be returned,
else the end point will be returned.}
}
\value{
A POINT geometry feature representing the first or last point in the
LINESTRING.
}
\description{
This is equivalent to ST_StartPoint and ST_EndPoint in PostGIS.
This is mainly used as a helper function for calculating trajectory locations.
}
\examples{
#start <- palms_trajectory_point(line, TRUE)
#end <- palms_trajectory_point(line, FALSE)

}
