% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palmsplus_shell.R
\name{palmsplus_shell}
\alias{palmsplus_shell}
\title{Shell function for processing palmsplus data}
\usage{
palmsplus_shell(
  palms_data,
  config_file,
  spatial_threshold = 200,
  temporal_threshold = 10,
  verbose = TRUE,
  days = TRUE,
  trajectories = TRUE,
  multimodal = TRUE,
  return_list = FALSE
)
}
\arguments{
\item{palms_data}{The palms dataset that is output from HABITUS. Path to csv file.}

\item{config_file}{The config file specifying the palmsplus processing parameters. Path to csv file.}

\item{spatial_threshold}{Spatial threshold for multimodal trips in meters. Default is \code{200}.}

\item{temporal_threshold}{Temporal threshold for multimodal trips in minutes. Default is \code{10}.}

\item{verbose}{Print progress after each step. Default is \code{TRUE}.}

\item{days}{Build the \code{days} dataset? Default is \code{TRUE}.}

\item{trajectories}{Build the \code{trajectories} dataset? Default is \code{TRUE}.}

\item{multimodal}{Build the \code{multimodal} dataset? Default is \code{TRUE}.}

\item{return_list}{Return all datasets as a list? Default is \code{FALSE}.}
}
\value{

}
\description{
Shell function for processing palmsplus data
}
