% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palmsplusr-package.R
\docType{package}
\name{palmsplusr}
\alias{palmsplusr}
\alias{palmsplusr-package}
\title{palmsplusr}
\description{
This package is a rewrite of palmsplus in r
}
\details{
In addition to the help pages, see the README page on \href{https://github.com/TheTS/palmsplusr}{github} for examples.
}
