% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{participant_basis}
\alias{participant_basis}
\title{Participant basis}
\format{
A data frame with 1 row and 3 variables:

\describe{
  \item{identifier}{Participant identifier}
  \item{school_id}{ID of school}
  \item{class_id}{ID of class}
}
}
\usage{
data(participant_basis)
}
\description{
A dataset containing
}
\keyword{datasets}
