% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_palms.R
\name{read_palms}
\alias{read_palms}
\title{Reading PALMS data from a csv file}
\usage{
read_palms(file)
}
\arguments{
\item{file}{Path to the file}
}
\value{
A sf dataframe of palms points.
}
\description{
Reading PALMS data from a csv file
}
\details{
This functions checks for the following columns, and will fail if
these are not present:
\itemize{
\item datetime
\item dow
\item lon
\item lat
\item fixtypecode
\item iov
\item tripnumber
\item triptype
\item tripmot
\item activity
\item activityintensity
\item activityboutnumber
\item sedentaryboutnumber
}
}
\examples{
#palms <- read_palms('palms_file.csv')

}
