% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{school}
\alias{school}
\title{School polygons}
\format{
A data frame with 1 row and 3 variables:

\describe{
  \item{school_id}{ID of school}
  \item{schoolname}{Name of school}
  \item{geometry}{Polygon geometry}
}
}
\usage{
data(school)
}
\description{
A dataset containing
}
\keyword{datasets}
