% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{palms_add_domain}
\alias{palms_add_domain}
\title{Adds an aggregation domain to the palmsplus_domains table.}
\usage{
palms_add_domain(name, formula)
}
\arguments{
\item{name}{The name of the domain as a string.}

\item{formula}{The formula as a string. The formula can contain variables
that are created with \code{palms_add_field}. Make sure to escape quotation
marks that are part of the formula.}
}
\value{
If the palmsplus_domains table is not present in the global environment,
it will be created. If it already exists, the new domain will be appended.
}
\description{
Adds an aggregation domain to the palmsplus_domains table.
}
\examples{
#palms_add_domain("d_home", "at_home")

}
