% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palms_build_days.R
\name{palms_build_days}
\alias{palms_build_days}
\title{Calculate day-level summaries from the palmsplus dataset}
\usage{
palms_build_days(data, config_file = NULL)
}
\arguments{
\item{data}{The palmsplus data obtained from \code{\link{palms_build_palmsplus}}.}

\item{config_file}{Path to the config file}
}
\value{
A table summarised by day.
}
\description{
Build a days dataset by summarising \code{palmsplus}
by day and person (\code{identifier}). Not all variables in \code{palmsplus}
are summarised, only those specified using \code{\link{palms_add_field}} with
the argument \code{domain_field = TRUE}. By default, a \code{duration} field
is added (e.g., the total minutes per day).

All data are summarised by default. However, additional aggragation \emph{domains}
can be specified using \code{\link{palms_add_domain}} before building days.
Domains are a subset of data, such as during school time. All \code{domain_field}
variables will be summarised for each \emph{domain} seperatly.
}
\examples{
data("palms")
palms_remove_tables()

# Add a field, and make it a domain_field
palms_add_field("mvpa", "activityintensity > 1", TRUE)

palmsplus <- palms_build_palmsplus(palms)

palms_add_domain("walking", "tripmot == 3")

# This will have 'walking' and 'total' domains
# Each domain will have 'duration' and 'mvpa' fields
days <- palms_build_days(palmsplus)

}
