% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multimodal.R
\name{palms_build_multimodal}
\alias{palms_build_multimodal}
\title{Build multimodal trips from trajectories}
\usage{
palms_build_multimodal(data, spatial_threshold, temporal_threshold,
  verbose = TRUE)
}
\arguments{
\item{data}{The trajectories object built with \code{palms_calc_trajectories}.}

\item{spatial_threshold}{Spatial threshold in meters}

\item{temporal_threshold}{Temporal threshold in minutes}

\item{verbose}{Print progress after each step. Default is \code{TRUE}.}
}
\value{
The input trajectories LINESTRING geometry, collapsed into multimodal trips
}
\description{
Build multimodal trips from trajectories.
}
\details{
Several columns are required in the \code{trajectories} dataset. These
need to be added as trajectory fields:
\itemize{
\item identifier
\item tripnumber
\item mot
\item start
\item end
\item geometry
}
}
\examples{
#multimodal <- palms_calc_multimodal(trajectories, 200, 10)

}
