% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palmsplus.R
\name{palms_calc_palmsplus}
\alias{palms_calc_palmsplus}
\title{Adds additional columns to PALMS data that represent domains and fields}
\usage{
palms_calc_palmsplus(data)
}
\arguments{
\item{data}{The palms data}
}
\value{
The input data with additional columns specified in
\code{palms_add_field} and \code{palms_add_domain}.
}
\description{
Adds additional columns to PALMS data that represent domains and fields
}
\examples{
#palmsplus <- palms_calc_palmsplus(palms)

}
