% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{palms_in_polygon}
\alias{palms_in_polygon}
\title{Check if points fall inside a polygon}
\usage{
palms_in_polygon(data, polygons, collapse_var = NULL)
}
\arguments{
\item{data}{The data points.}

\item{polygons}{The polygon feature.}

\item{collapse_var}{An optional parameter for aggragating polygons by a
variable. Essentially this dissolves the polygon layer by \code{collapse_var}.}
}
\value{
A logical vector the length of the input data points. If multiple
polygons are present, an additional column will be returned for each.
}
\description{
Returns a logical vector the length of the input \code{data}
points. This checks if each point falls inside \code{polygon}. Multiple
polygons are supported (e.g., multiple homes, or parks). If the \code{polygon}
data is empty, \code{NA} will be returned.
}
\examples{
data(list = c("palms", "home"))

home.buffer <- palms_buffer(home, distance = 100)

at_home <- palms_in_polygon(palms, home.buffer)

}
