% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{palms_in_time}
\alias{palms_in_time}
\title{Returns a logical vector if points are within a time period}
\usage{
palms_in_time(data, pid, timetable, basis, start_col, end_col)
}
\arguments{
\item{data}{The data points}

\item{pid}{The participant identifier}

\item{timetable}{The class timetable table}

\item{start_col}{The column in \code{timetable} containing period start times}

\item{end_col}{The column in \code{timetable} containing period end times}

\item{participant_basis}{The participant basis table}
}
\value{
A logical vector the length of the input data points.
}
\description{
Returns a logical vector if points are within a time period
}
\examples{
#data <- ...
#ct <- read.csv("class_timetable.csv")
#pb <- read.csv("participant_basis.csv")
#data$in_recess <- palms_in_time(data, "BC0627", ct, pb, "recess_start", "recess_end")

}
