% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palmsplus_shell.R
\name{palmsplus_shell}
\alias{palmsplus_shell}
\title{Shell function for processing palmsplus data}
\usage{
palmsplus_shell(
  palms_data,
  habitus = FALSE,
  config_file,
  spatial_threshold = 200,
  temporal_threshold = 10,
  verbose = TRUE,
  save_output = TRUE,
  output_path = getwd(),
  return_list = FALSE,
  days = TRUE,
  trajectories = TRUE,
  multimodal = TRUE
)
}
\arguments{
\item{palms_data}{The palms dataset that is output from HABITUS. Path to csv file.}

\item{habitus}{Is this data from HABITUS (as opposed to PALMS, which has different column names). Default is \code{FALSE}.}

\item{config_file}{The config file specifying the palmsplus processing parameters. Path to csv file.}

\item{spatial_threshold}{Spatial threshold for multimodal trips in meters. Default is \code{200}.}

\item{temporal_threshold}{Temporal threshold for multimodal trips in minutes. Default is \code{10}.}

\item{verbose}{Print progress after each step. Default is \code{TRUE}.}

\item{save_output}{Logical. Save output files to disk. Default is \code{TRUE}.}

\item{output_path}{Path to directory where output files are saved. Default is current working directory}

\item{return_list}{Return all datasets as a list? Default is \code{FALSE}.}

\item{days}{Build the \code{days} dataset? Default is \code{TRUE}.}

\item{trajectories}{Build the \code{trajectories} dataset? Default is \code{TRUE}.}

\item{multimodal}{Build the \code{multimodal} dataset? Default is \code{TRUE}.}
}
\value{
Returns a list with requested data frames (if `return_list` is true)
}
\description{
Shell function for processing palmsplus data
}
\examples{
\dontrun{
palmsplus_shell(palms_data = "habitus_cleaned.csv", # Path to palms/habitus dataset
                habitus = TRUE,                     # Are data from habitus (not palms)

                config_file = "config.csv",         # Path to config file

                days = TRUE,                        # Build the 'days' dataset?
                trajectories = TRUE,                # Build the 'trajectories' dataset?
                multimodal = TRUE,                  # Build the 'multimodal' dataset?

                spatial_threshold = 100,            # Spatial threshold for multimodal trips (meters)
                temporal_threshold = 5,             # Temporal threshold for multimodal trips (minutes)

                save_output = TRUE,                 # Save output files?
                output_path = "output",             # Directory where results are saved
                return_list = TRUE)                 # Return output files as a list




}


}
